+++ mobile/RELEASES.md
@@ -65,8 +65,7 @@ Der Build-Zusatz kommt aus `expo.android.versionCode`. Dadurch ist fuer Tester k
 Historische Einordnung, neueste Version oben:
 
 ```text
-1.0.0
-0.1.0-beta.1
+0.3.0-alpha.1
 0.2.0-alpha.1
 0.1.1-alpha.1
 0.1.0-alpha.14
@@ -90,7 +89,7 @@ Historische Einordnung, neueste Version oben:
 In GitHub unter `Actions` den Workflow `Mobile APK Release` starten. Der Workflow erzeugt ein Artefakt mit diesem Namensschema:
 
 ```text
-LionsAPP-ALPHA-v0.1.1-build15-<commit>.apk
+LionsAPP-ALPHA-v0.3.0-build17-<commit>.apk
 ```
 
 ## Release-Signatur